move the IRQ handler into main.c
