+++ driver/bluetooth_low_energy.h
@@ -28,8 +28,25 @@
 
 #define BLE_MAX_MTU_LENGTH 128
 
+// Buffer sizes for REPL ring buffers; +45 allows a bytearray to be printed in one go.
+#define RING_BUFFER_LENGTH (1024 + 45)
+
+/**
+ * @brief Ring buffers for the repl rx and tx data which goes over BLE.
+ */
+typedef struct
+{
+    uint8_t buffer[RING_BUFFER_LENGTH];
+    uint16_t head;
+    uint16_t tail;
+} ring_buf_t;
+
 extern uint16_t ble_negotiated_mtu;
 
+extern ring_buf_t nus_rx;
+extern ring_buf_t nus_tx;
+extern uint16_t ble_conn_handle;
+extern uint8_t ble_adv_handle;
 extern void ble_on_connect(void);
 extern void ble_on_disconnect(void);
 
@@ -38,3 +55,9 @@ void ble_nus_tx(char const *buf, size_t len);
 void ble_raw_tx(uint8_t const *buf, uint16_t len);
 int ble_nus_rx(void);
 bool ble_nus_is_rx_pending(void);
+
+// wrapper over ring buffer operations;
+bool ring_full(ring_buf_t const *ring);
+bool ring_empty(ring_buf_t const *ring);
+void ring_push(ring_buf_t *ring, uint8_t byte);
+uint8_t ring_pop(ring_buf_t *ring);
